+++ src/features/auth/auth.seller.service.spec.ts
@@ -28,7 +28,6 @@ import {
   SELLER_CREDENTIAL_REPOSITORY,
   type ISellerCredentialRepository,
 } from '@/features/auth/repositories/seller-credential.repository.interface';
-import { OidcClientService } from '@/features/auth/services/oidc-client.service';
 import { TokenService } from '@/features/auth/services/token.service';
 import { TOKEN_SERVICE } from '@/features/auth/services/token.service.interface';
 
@@ -91,7 +90,6 @@ describe('AuthService (seller)', () => {
           provide: JwtService,
           useValue: { sign: jest.fn(() => 'mock-access-token') },
         },
-        { provide: OidcClientService, useValue: {} },
         {
           provide: TOKEN_SERVICE,
           useClass: TokenService,
